# [Merged by Bors] - Shrink DrawFunctionId