--- a/crates/bevy_render/src/lib.rs
+++ b/crates/bevy_render/src/lib.rs
@@ -1,5 +1,5 @@
 #[cfg(target_pointer_width = "16")]
-compile_error!("bevy_render cannot safely compile for a 16-bit platform.");
+compile_error!("bevy_render cannot compile for a 16-bit platform.");
 
 extern crate core;